pyproject.toml
[project]
name = "test-xradar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arm-pyart>=2.1.1",
    "jupyterlab>=4.5.1",
    "kerchunk>=0.2.9",
    "xradar>=0.11.0",
    "virtualizarr>=2.2.1",
    "zarr>=3.1.5",
    "numpy<2.4.0",
    "virtualizarr",
]

